# docs/swagger.yaml
basePath: /api
definitions:
...
  api.SignUpRequest:
    properties:
      avatarURL:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      lastName:
        type: string
      location:
        type: string
      organizationID:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      provider:
        type: string
      providerID:
        type: string
      username:
        type: string
    type: object
...
  api.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      gender:
        type: string
      is_active:
        type: boolean
      is_superuser:
        type: boolean
      last_name:
        type: string
      locale:
        type: string
      phone_number:
        type: string
      phone_verified:
        type: boolean
      username:
        type: string
    type: object
  api.UpdateUserSelfRequest:
    properties:
      avatar_url:
        type: string
      date_of_birth:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      locale:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
...
  httpresp.HTTPError:
    properties:
      code:
        description: the HTTP status
        example: 500
        type: integer
      error:
        description: the error description
        example: internal server error
        type: string
      success:
        description: wether the business action is successfull or not
        example: false
        type: boolean
    type: object
...
  interfaces.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      date_of_birth:
        description: Using string pointer for optional date
        type: string
      email:
        description: Authentication
        type: string
      email_verified:
        description: Account status
        type: boolean
      first_name:
        type: string
      gender:
        type: string
      id:
        description: Primary key
        type: string
      is_active:
        type: boolean
      is_superuser:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      last_password_change:
        type: string
      locale:
        type: string
      metadata:
        additionalProperties: true
        description: Metadata
        type: object
      mfa_enabled:
        description: Security
        type: boolean
      phone_number:
        type: string
      phone_verified:
        type: boolean
      provider:
        type: string
      provider_id:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        description: User profile
        type: string
    type: object
...
host: localhost:5001
info:
  contact:
    email: malek.radhouen@gmail.com
    name: API Support
    url: https://www.magic.fr/support
  description: Here is our solution documentation and testing portal of provided functionalities
    to interact with our hypervisor tool.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.magic.com/terms/
  title: Magic - Inventory solution API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/interfaces.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/interfaces.User'
        "400":
          description: Data validation error
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a user by ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            $ref: '#/definitions/interfaces.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Get a user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update profile fields of the authenticated user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api.UpdateUserSelfRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/interfaces.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Update a user (self)
      tags:
      - users
  /users/{id}/admin:
    patch:
      consumes:
      - application/json
      description: Admin-only update that may flip privileged fields
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Privileged user update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/api.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/interfaces.User'
      summary: Update a user (admin)
      tags:
      - users
  /users/admins:
    get:
      consumes:
      - application/json
      description: Get all users with admin, root or superadmin role
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of admin users
          schema:
            items:
              $ref: '#/definitions/interfaces.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Get admin users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with admin role (admin-only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Admin user information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Admin user created
          schema:
            $ref: '#/definitions/interfaces.User'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create an admin user
      tags:
      - users
  /users/customers:
    get:
      consumes:
      - application/json
      description: Get all users with the user role (customers)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of customers
          schema:
            items:
              $ref: '#/definitions/interfaces.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpresp.HTTPError'
      summary: Get customer users
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...